requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.2
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
drf-yasg==1.17.0
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyparsing==2.4.6
python-decouple==3.3
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
